requirements.txt
matplotlib==3.10.0
numpy==1.23.5
pandas==2.2.3
Pillow==11.1.0
prettytable==3.12.0
pytorch_lightning==2.4.0
rasterio==1.3.11
Requests==2.32.3
scikit_learn==1.6.0
seaborn==0.13.2
tensorboard==2.18.0
timm==1.0.12
torch==2.4.1
torchmetrics==1.4.2
torchsummary==1.5.1
torchvision==0.19.1
torchviz==0.0.2
tqdm==4.66.5
zstandard==0.23.0
pyarrow==17.0.0
fastparquet==2024.2.0
torchcam == 0.4.0
